// src/shared/services/api/models/forumMessageDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export interface ForumMessageDto {
  id?: string;
  course_id?: string;
  user_id?: string;
  /** @nullable */
  user_full_name?: string | null;
  /** @nullable */
  message_text?: string | null;
  date_created?: string;
}