// src/shared/services/api/models/postApiV1CourseContentBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export type PostApiV1CourseContentBody = {
  course_id?: string;
  meta_topic?: string;
  meta_topic_order?: number;
  meta_duration_minute?: number;
  title?: string;
  content_type_id?: string;
  video_url?: string;
  pdf_file?: Blob;
  sequence_order?: number;
};