// src/shared/services/api/models/putApiV1CourseCourseIdBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export type PutApiV1CourseCourseIdBody = {
  title?: string;
  description?: string;
  category_id?: string;
  sub_category_id?: string;
  difficulty_level_id?: string;
  duration_minute?: number;
  thumbnail_img?: Blob;
  is_drafted?: boolean;
};